+++ CHANGELOG.md
@@ -46,6 +46,8 @@ Standardise files with files in sous-chefs/repo-management
 
 Standardise files with files in sous-chefs/repo-management
 
+- Compensate for NPM CLI deprecating `-global` and use `--location=global
+
 ## 10.1.0 - *2022-08-08*
 
 - To ensure consistent environment between `npm install` and `npm list`, pass same environment variables.